googlecloud/firestore/promotions/requirements.txt
Flask
flask-cors
google-cloud-firestore
google-cloud-logging